correct imports, separate sim-settings for EP and comfort plugin
